scx: separate out the user_exit_info BPF and user space code
We need to separate out the user_exit_info bpf code and user space code
in order to make it easier to enable unit testing of the BPF code.
